Running gulp responsive-images now runs about 100x faster.

```
gulp responsive-images  108.11s user 3.93s system 166% cpu 1:07.48 total
```

Now:
```
gulp responsive-images  5.82s user 1.79s system 68% cpu 11.167 total
```

Much :heart: to @callumacrae :tada:  for figuring it out!!

The issue was ordering of the pipes. We need to make sure the files are renamed, then checked for change before handling resize or optimisation

:tada: :tada: :tada: :tada: :tada: :tada: :tada: