@callumacrae 2015-08-10
Little issue (also mentioned on Slack): images generated from @2x images will be as twice as big as they need to be as the check for `@2x` in the filename will always fail. The way to fix that is to save the old file name before renaming it and calling gm directly, looking for the new property instead of `file.path`.