@callumacrae 2015-08-11
If the test failure was the same as on master, rebasing onto master should fix it.