@callumacrae 2015-08-13
Rerunning the tests then merging and releasing :+1: