@onlymejosh 2015-08-13
Im just about to push this to staging to test :)

On Thu, Aug 13, 2015 at 10:34 AM, Callum Macrae <notifications@github.com>
wrote:

> Rerunning the tests then merging and releasing :+1:
> ---
> Reply to this email directly or view it on GitHub:
> https://github.com/Lostmyname/lmn-gulp-tasks/pull/19#issuecomment-130591662